ws: publisher: Update "ws" api type flag with "type" flag
There can be multiple API types in the future. Keeping a flag for
each API type is not useful. Therefore "ws" API type flag was updated
with "type" flag. Whenever new API type is implemented it has to be
introduced to APIConstants.APIType enum. If type is not provided type
will be default to "HTTP".