Remove API Product view from store
It was discussed that the API Product is identical in functionality to a standard API from the perspective of the developer portal/API Subscribers.
The differentiation between the API Product and the API is only evident and required by the API Publisher/Creator. Therefore the API Product will be
displayed along with standard APIs in the same views already available in the developer portal. This commits removes all API product specific
views/functionality from the developer portal.