Change endpoint update error log to debug log
With the introduction of API Products it is possible that an APIs endpoint could be already deployed prior
to an API being published or API is published prior to creating of the API Product. Therefore Endpoints will
be updated by default at all times. In case the endpoint has not been created already, removal of the endpoint
prior to updating cannot be done. Since this is part of the possible behaviour now, this error log has been
changed to a debug log.