Improve Application details extraction for recommendation feature
having two different methods to publish new applications and update
applications with different parameters will create errors in
recommendation server.
when a “application update” request sent for an application which was
not recorded in the server will create an error. If the recommendation
option is enabled after a while, this error will be frequent and there is
no other option to get application details. new changes will have only
one method to publish application details and in the server side, if a
previous record does not exists, it will record the details as a new
application.

Signed-off-by: Lakshitha Gunasekara <lakshitha.gunasekara@gmail.com>