Add action attrib to API_UPDATE event
CC only process deployment events. However, when default version
is updated it needs to process the API_UPDATE event to redeploy
the API. Redeploying API for all API_UPDATE events is not an
option for CC. Therefore, to identify the purpose of API_UPDATE
event we are adding this `action` attribute. With that we can
inform event receiving parties what type of change has happened.