Fix API update when publishing while deprecating old versions
There is an option at the lifecycle change to publish an API while
deprecating other versions of the same API. Here, API object of
other API are updated with the deprecated state, however the
API object is having very minimal information. This breaks the
newly deprecated API in the UI portals as it does not contain the
minimum information needed to show in publisher and dev portal
ui.

Resolves: https://github.com/wso2/api-manager/issues/1115