Added sandbox key capability to publisher tryout console
The publisher portal tryout console should work for either production or sandbox endpoint. But if there is no production endpoint, we could not invoke the API.

This is due to our current code only having the capability to handle this only if a production endpoint is available. In this commit, I added sandbox key capability to publisher tryout console to fix https://github.com/wso2/api-manager/issues/1190