Add the support to generate SOAP to REST operation elements in order
In the current SOAP to REST API creation, the SOAP
envelop elements are not mapped in the same order
defined in the wsdl. Current implementation used JSONObjects
to record the parameters and this will not preserve the order.
However if the operation has sequence elements, changed
order will break the operation. Instead of JSONObject,
this implementation uses the swagger example data type
which internally use Linked Hash maps.

Resolves: https://github.com/wso2/api-manager/issues/1441